Moved filter checkboxes below the search bar and added tags as an additional filter
